Shifted everything in a subfolder, so it's easier to install.
