+++ py-comment-counter/hooks/pre-commit
@@ -1,0 +1,64 @@
+#!/usr/bin/env python
+import configparser
+import os
+import git
+from subprocess import check_output
+from git import Repo
+
+# set the variable with the current directory
+dir_path = os.path.abspath(os.curdir)
+# parser of .ini file
+config = configparser.ConfigParser()
+# build path for conf.ini file
+# path_conf_file = os.path.join(dir_path, os.pardir)
+path_conf_file = os.path.join('py-comment-counter', 'config_script',
+                              'config.ini')
+config.read(path_conf_file)
+percentage_min_comments = config.getint('Comments_Counter', 'percentage_min_comments')
+
+def _cut_useles(files):
+    """
+    Cut the useles part of a raw and take only the file to commit
+    :param files: list of files to commit or not
+    :return file_list: list with only file to commit
+    """
+    files_list = []
+    files = files.split('\n')
+    for file in files:
+        if not file.startswith('?') and not file.startswith(' ') and \
+           file.endswith('.py'):
+            file = file[2:]
+            file = file.strip(' ')
+            files_list.append(file)
+    return files_list
+
+# initialize the repository variable
+repo = Repo(dir_path)
+git = repo.git
+# run the command git status -s
+status = git.status('-s')
+status = _cut_useles(status)
+# build path for Main.py
+program_path = os.path.join(dir_path, 'py-comment-counter', 'Main.py')
+for file_name in status:
+    # passing the file for the count of comments
+    file = os.path.join(dir_path, file_name)
+    perc = check_output(['python', program_path, file])
+    perc = perc.decode()
+    perc = perc.split('\r')
+    if perc[0] == 'pydoc assente':
+        print("Nel file: " + file_name +
+              ' pydoc mancanti o non sempre presenti')
+        exit(1)
+    if perc[0] == 'file empty':
+        print("Il file: " + file_name +
+              ' non presenta niente al suo interno')
+    if float(perc[0]) < percentage_min_comments:
+        print('Nel file: ' + file_name + ' i commenti sono meno del %s' %
+              percentage_min_comments)
+        print('Percentuale commenti: ' + perc[0])
+        exit(1)
+    else:
+        print('Nel file: ' + file_name + ' hai commentato bene, bravo')
+        print('Percentuale commenti: ' + perc[0] + '%')
+exit(0)
